Updates to .page files and grunt file. Addition of new resource.
Updated the grunt file with a then block in the prompt task so that
the one prompt task will work properly regardless of the deploy target

Installed the bootstrap resource and renamed it from LABJS_Bootstrap to
Bootstrap.

Updated the .page files that reference the component and the static
resources to point to the new names.